// models/enums/OpportunityType.ts
export enum OpportunityType {
    FARMING = 'FARMING',                 // 農業工作
    GARDENING = 'GARDENING',             // 園藝工作
    ANIMAL_CARE = 'ANIMAL_CARE',         // 動物照顧
    CONSTRUCTION = 'CONSTRUCTION',       // 建築工作
    HOSPITALITY = 'HOSPITALITY',         // 接待服務
    COOKING = 'COOKING',                 // 烹飪工作
    CLEANING = 'CLEANING',               // 清潔工作
    CHILDCARE = 'CHILDCARE',             // 兒童照顧
    ELDERLY_CARE = 'ELDERLY_CARE',       // 老人照顧
    TEACHING = 'TEACHING',               // 教學工作
    LANGUAGE_EXCHANGE = 'LANGUAGE_EXCHANGE', // 語言交換
    CREATIVE = 'CREATIVE',               // 創意工作
    DIGITAL_NOMAD = 'DIGITAL_NOMAD',     // 數位遊牧
    ADMINISTRATION = 'ADMINISTRATION',   // 行政工作
    MAINTENANCE = 'MAINTENANCE',         // 維修工作
    TOURISM = 'TOURISM',                 // 旅遊工作
    CONSERVATION = 'CONSERVATION',       // 保育工作
    COMMUNITY = 'COMMUNITY',             // 社區工作
    EVENT = 'EVENT',                     // 活動工作
    OTHER = 'OTHER'                      // 其他
  }