// models/enums/WorkType.ts
export enum WorkType {
  RECEPTION = 'reception',           // 接待/櫃檯
  HOUSEKEEPING = 'housekeeping',     // 房務/清潔
  FOOD_SERVICE = 'food_service',     // 餐飲服務
  KITCHEN = 'kitchen',               // 廚房助理
  FARMING = 'farming',               // 農場工作
  GARDENING = 'gardening',           // 園藝/花藝
  CONSTRUCTION = 'construction',     // 建築/維修
  TEACHING = 'teaching',             // 教學/輔導
  TOUR_GUIDE = 'tour_guide',         // 導遊/嚮導
  ANIMAL_CARE = 'animal_care',       // 動物照顧
  MARKETING = 'marketing',           // 行銷/社群媒體
  PHOTOGRAPHY = 'photography',       // 攝影/影片製作
  TRANSLATION = 'translation',       // 翻譯/語言協助
  ADMINISTRATION = 'administration', // 行政/辦公室
  IT_SUPPORT = 'it_support',         // IT/技術支援
  CREATIVE = 'creative',             // 創意/藝術
  EVENT = 'event',                   // 活動策劃/執行
  OTHER = 'other'                    // 其他
}